+++ src/pages/Plugins/components/pluginViews/Notion.vue
@@ -10,77 +10,86 @@
           ><i-mdi-connection />&nbsp;{{ t("checkState") }}</x-button
         >
       </div>
-      <div
-        class="mb-3 flex h-26 w-full items-center justify-center rounded-md border border-gray-200 bg-white p-3"
+
+      <x-message
+        :type="workspaceInfo.id ? 'info' : 'warning'"
+        :dismissible="false"
       >
-        <div v-if="workspaceInfo.id" class="flex h-full w-full gap-3">
-          <img :src="workspaceInfo.avatar_url" alt="notion" class="h-20 w-20" />
-          <div class="flex-1">
-            <div class="ml-2 text-xl leading-6 font-semibold text-gray-700">
-              Bot: {{ workspaceInfo.name }}
-            </div>
-            <div class="text-md ml-2 leading-6 text-gray-400">
-              Work Space:&nbsp;{{ workspaceInfo.workspace_name }}
-            </div>
-            <div class="text-primary ml-2 text-sm leading-6">
-              UIChat 已成功配置到 Notion！
-            </div>
-          </div>
-        </div>
-        <x-spin v-else-if="!workspaceInfo.id && loading" />
-        <x-result v-else class="w-[30%]" type="404" />
-      </div>
+        {{ workspaceInfo.id ? "UIChat 已配置notion" : "请检查链接" }}
+        <x-button
+          size="small"
+          v-if="workspaceInfo.id"
+          @click="(e) => (visible = true)"
+          type="ghost"
+          class="ml-4"
+          >写文章</x-button
+        >
+      </x-message>
 
       <div class="w-full">
-        <x-input
-          :disabled="loading || !workspaceInfo.id"
-          v-model="testArticleTitle"
-          class="mb-3 w-full"
-          placeholder="给Notion发送一段测试知识"
-        />
-        <x-textarea
-          :disabled="loading || !workspaceInfo.id"
-          class="h-full"
-          :rows="28"
-          v-model="articleContent"
-          placeholder="给Notion发送一段测试知识"
-        />
-      </div>
-    </template>
-    <template #pluginConsoleOperation>
-      <x-button
-        :disabled="loading || !workspaceInfo.id"
-        @click="execute"
-        class="w-full"
-        >创建一个新的Notion Page</x-button
-      >
-    </template>
-
-    <template #pluginDebugger>
-      <div class="flex h-full w-full flex-col gap-0">
         <x-list
           :title="'Page列表'"
           @select="onSelectPage"
           :items="pageList"
-          :class="{
-            'h-[25%]': !!pageContent,
-            'h-full': !pageContent,
-          }"
-          class="mb-3 w-full"
+          class="my-3 h-[630px] w-full"
         />
+        <!-- <x-input
+          :disabled="loading || !workspaceInfo.id"
+          v-model="testArticleTitle"
+          class="mb-3 w-full"
+          placeholder="标题"
+        />
+        <x-textarea
+          :disabled="loading || !workspaceInfo.id"
+          :rows="24"
+          v-model="articleContent"
+          placeholder="给Notion发送一段测试知识"
+        /> -->
+      </div>
+    </template>
+    <template #pluginConsoleOperation>
+      <!-- <x-button
+        :disabled="loading || !workspaceInfo.id"
+        @click="execute"
+        class="w-full"
+        >创建一个新的Notion Page</x-button
+      > -->
+    </template>
 
+    <template #pluginDebugger>
+      <div class="flex h-full w-full flex-col gap-0">
         <div
-          :class="{
-            'flex-1 p-10': !!pageContent,
-            'h-0': !pageContent,
-          }"
-          class="w-full max-w-full overflow-auto rounded-md border border-gray-300 transition-all duration-300 ease-in-out"
+          class="w-full max-w-full flex-1 overflow-auto rounded-md border border-gray-300 p-10 transition-all duration-300 ease-in-out"
         >
-          <x-markdown class="max-w-[790px]" :content="pageContent" />
+          <x-spin v-if="loading" class="mt-[25%]" />
+          <x-markdown
+            v-if="!loading && pageContent"
+            class="max-w-[790px]"
+            :content="pageContent"
+          />
+          <x-result v-if="!loading && !pageContent" type="404" />
         </div>
       </div>
     </template>
   </plugin-page-wrapper>
+
+  <x-dialog :visible="visible" @cancel="visible = false" @confirm="execute">
+    <template #header>写一篇Page</template>
+    <template #body>
+      <x-input
+        :disabled="loading || !workspaceInfo.id"
+        v-model="testArticleTitle"
+        class="mb-3 w-full"
+        placeholder="标题"
+      />
+      <x-textarea
+        :disabled="loading || !workspaceInfo.id"
+        :rows="24"
+        v-model="articleContent"
+        placeholder="给Notion发送一段测试知识"
+      />
+    </template>
+  </x-dialog>
 </template>
 
 <script setup>
@@ -110,6 +119,7 @@ const pageList = ref([]);
 const pageContent = ref("");
 const { t } = useNamespace("PluginsViews.Notion");
 
+const visible = ref(false);
 const setLoading = (value) => {
   loading.value = value;
 };
@@ -151,10 +161,12 @@ const execute = async () => {
     console.error(error);
   } finally {
     setLoading(false);
+    visible.value = false;
   }
 };
 
 const onSelectPage = async (item) => {
+  setLoading(true);
   try {
     pageContent.value = "";
     const { id } = item;
@@ -162,6 +174,8 @@ const onSelectPage = async (item) => {
     pageContent.value = pageData;
   } catch (error) {
     console.error(error);
+  } finally {
+    setLoading(false);
   }
 };
 </script>