+++ src/plugins/cachePlugin/schema.ts
@@ -52,15 +52,10 @@ const providersMap: ConfigMap = {
     provider: "cloudflare",
     baseURL: `${BASE_URL}cloudflare`,
   },
-
-  // qwen: {
-  //   provider: "qwen",
-  //   baseURL: "http://localhost:5090/qwen/v1",
-  //   apiKey: "",
-  //   model: "qwen-v1-8k",
-  //   systemPrompt:
-  //     "你是 Qwen，由 Qwen AI 提供的人工智能助手，你更擅长中文和英文的对话。你会为用户提供安全，有帮助，准确的回答。同时，你会拒绝一切涉及恐怖主义，种族歧视，黄色暴力等问题的回答。Qwen AI 为专有名词，不可翻译成其他语言。",
-  // },
+  openrouter: {
+    provider: "openrouter",
+    baseURL: `${BASE_URL}openrouter`,
+  },
 };
 
 const conversationsList: ConversationType[] = [];